requirements.txt
# Dependencias del proyecto
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0

